chore: move hooks to it's own page
